+++ elements/wifi/ap/proberesponder.cc
@@ -308,7 +308,7 @@ ProbeResponder_write_param(const String &in_s, Element *e, void *vparam,
 {
   ProbeResponder *f = (ProbeResponder *)e;
   String s = cp_uncomment(in_s);
-  switch((int)vparam) {
+  switch((intptr_t)vparam) {
   case H_DEBUG: {    //debug
     bool debug;
     if (!cp_bool(s, &debug))
