+++ elements/wifi/sr/gatewayselector.cc
@@ -425,7 +425,7 @@ String
 GatewaySelector::read_param(Element *e, void *vparam)
 {
   GatewaySelector *f = (GatewaySelector *) e;
-  switch ((int)vparam) {
+  switch ((intptr_t)vparam) {
   case 0:			
     return String(f->_is_gw) + "\n";
   case 1:			
@@ -460,7 +460,7 @@ GatewaySelector::write_param(const String &in_s, Element *e, void *vparam,
 {
   GatewaySelector *f = (GatewaySelector *)e;
   String s = cp_uncomment(in_s);
-  switch((int)vparam) {
+  switch((intptr_t)vparam) {
   case 0: {    //is_gateway
     bool b;
     if (!cp_bool(s, &b))
