+++ elements/wifi/sr/srqueryresponder.cc
@@ -370,7 +370,7 @@ SRQueryResponder_write_param(const String &in_s, Element *e, void *vparam,
 {
   SRQueryResponder *f = (SRQueryResponder *)e;
   String s = cp_uncomment(in_s);
-  switch((int)vparam) {
+  switch((intptr_t)vparam) {
   case H_DEBUG: {    //debug
     bool debug;
     if (!cp_bool(s, &debug))
