+++ elements/wifi/sr2/sr2queryforwarder.cc
@@ -364,7 +364,7 @@ SR2QueryForwarder_write_param(const String &in_s, Element *e, void *vparam,
 {
   SR2QueryForwarder *f = (SR2QueryForwarder *)e;
   String s = cp_uncomment(in_s);
-  switch((int)vparam) {
+  switch((intptr_t)vparam) {
   case H_DEBUG: {    //debug
     bool debug;
     if (!cp_bool(s, &debug))
